otimizando_sistema_bancario

+++ Desafio_otimizando_sistema_bancario.py
@@ -172,7 +172,7 @@ def main():
         elif opcao == "r":
             lista_operadoras()
             input("Informe o número de telefone (somente números): ")
-            valor = float(input("Informe o valor da recarga: "))
+            valor = int(input("Informe o valor da recarga: "))
 
 
             saldo, extrato = recarregar(
